Factored all code into h264 decoder and handled encoding through input variable
