--- crates/utils/re_video/src/decode/ffmpeg_h265/mod.rs
@@ -1,4 +1,0 @@
-mod ffmpeg;
-mod version;
-
-pub use ffmpeg::FFmpegCliH265Decoder;
