+++ crates/utils/re_video/src/h265.rs
@@ -5,7 +5,6 @@ use cros_codecs::codec::h265::parser::Sps;
 
 /// Retrieve [`VideoEncodingDetails`] from an H.265 SPS.
 pub fn encoding_details_from_h265_sps(sps: &Sps) -> VideoEncodingDetails {
-    // 1) Build the codec string from profile and level:
     let profile_idc = sps.profile_tier_level.general_profile_idc;
     let level_idc: u8 = sps.profile_tier_level.general_level_idc as u8;
     // WebCodecs HEVC strings are usually "hvc1.<profile>.<level>" (with optional flags)
@@ -15,15 +14,12 @@ pub fn encoding_details_from_h265_sps(sps: &Sps) -> VideoEncodingDetails {
         level = level_idc
     );
 
-    // 2) Dimensions (accounts for any conformance window):
     let width = sps.width();
     let height = sps.height();
     let coded_dimensions = [width, height];
 
-    // 3) Bit depth = luma_minus8 + 8
     let bit_depth = Some(sps.bit_depth_luma_minus8 + 8);
 
-    // 4) Chroma subsampling from chroma_format_idc: 0=mono,1=4:2:0,2=4:2:2,3=4:4:4
     let chroma_subsampling = match sps.chroma_format_idc {
         0 => Some(ChromaSubsamplingModes::Monochrome),
         1 => Some(ChromaSubsamplingModes::Yuv420),
@@ -37,6 +33,6 @@ pub fn encoding_details_from_h265_sps(sps: &Sps) -> VideoEncodingDetails {
         coded_dimensions,
         bit_depth,
         chroma_subsampling,
-        stsd: None, // Annex B streams never include an MP4 stsd box
+        stsd: None,
     }
 }
